Make expire a 32 bit integer... 16 might be too short in some case.
git-svn-id: file:///home/yoann/SVN/trunk/libprelude@3876 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/include/timer.h
@@ -31,7 +31,7 @@
 typedef struct {
         prelude_list_t list;
 
-        short int expire;
+        int expire;
         time_t start_time;
 
         void *data;
